Organize Falcon NTRU tables and modp tests
